# swagger-admin.yml
...
swagger: "2.0"
info:
  version: "2016-12-20T18:27:47Z"
  title: "b2b-admin-apigateway"
basePath: "/"
schemes:
  - "https"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
  auth0_jwt:
    type: apiKey
    name: Authorization
    in: header
    x-amazon-apigateway-authtype: custom
    x-amazon-apigateway-authorizer:
      identitySource: method.request.header.Authorization
      authorizerResultTtlInSeconds: 0
      authorizerUri:
        Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${JWTAuthorizerLambdaFunction.Arn}/invocations
      authorizerCredentials:
        Fn::Sub: arn:aws:iam::139820419717:role/${ApiGatewayIamRole}
      type: token
paths:
...
  /reseller/{resellerId}:
    post:
      operationId: updateReseller
      produces:
        - "application/json"
      parameters:
        - name: resellerId
          in: path
          description: Identifier of reseller ID
          required: true
          type: string
        - name: somePayLoad
          in: body
          description: requestBody payload for update reseller
          schema:
            $ref: '#/definitions/Reseller'
      responses:
        200:
          description: "200 response"
          headers:
            Content-Type:
              type: "string"
      security:
        - auth0_jwt: [ ]
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: "arn:aws:iam::139820419717:role/${ApiGatewayIamRole}"
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Content-Type: "'application/json'"
        # NOTE: Replace <<region>> and <<account>> fields
        uri:
          Fn::Sub: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:139820419717:function:${AdminHandlerLambdaFunction}/invocations
        passthroughBehavior: "when_no_match"
        httpMethod: "POST"
        type: "aws_proxy"
    get:
      operationId: getResellerByExternalId
      produces:
        - "application/json"
      parameters:
        - name: resellerId
          in: path
          description: Identifier of employer
          required: true
          type: string
      responses:
        200:
          description: "200 response"
          headers:
            Content-Type:
              type: "string"
      security:
        - auth0_jwt: []
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: "arn:aws:iam::139820419717:role/${ApiGatewayIamRole}"
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Content-Type: "'application/json'"
        # NOTE: Replace <<region>> and <<account>> fields
        uri:
          Fn::Sub: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:139820419717:function:${AdminHandlerLambdaFunction}/invocations
        passthroughBehavior: "when_no_match"
        httpMethod: "POST"
        type: "aws_proxy"
...
definitions:
  Reseller:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      eligibility_rules:
        type: object
        properties:
          productTypes:
            type: array
            items:
              type: string
          validationFields:
            type: array
            items:
              type: string
      configurations:
        type: object
        properties:
          sku_mappings:
            type: object
          sso:
            type: object
            properties:
              cert_filename:
                type: string
          autoAssignPCP :
            type : string
      support_phone:
        type: string
      support_email:
        type: string
      b2b_link:
        type: string
      kickoff_link:
        type: string
      epic_link:
        type: string
...